requirements.txt
asgiref==3.6.0
attrs==23.1.0
autopep8==2.0.2
black==23.3.0
Brotli==1.0.9
click==8.1.3
colorama==0.4.6
dj-database-url==2.0.0
Django==4.2
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
gprof2dot==2022.7.29
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.4.0
